backend/requirements.txt
fastapi
uvicorn
cloud-sql-python-connector[pg8000]
pg8000
sqlalchemy
pydantic
google-cloud-aiplatform
vertexai
